+++ aeron-client/src/test/java/uk/co/real_logic/aeron/logbuffer/TermAppenderTest.java
@@ -96,7 +96,7 @@ public class TermAppenderTest
 
     private int frameLengthAligned(final int frameLength)
     {
-        return (int) align(frameLength, FRAME_ALIGNMENT);
+        return align(frameLength, FRAME_ALIGNMENT);
     }
 
     private int capacity()
