+++ aeron-driver/src/test/java/uk/co/real_logic/aeron/driver/LossDetectorTest.java
@@ -47,7 +47,7 @@ public class LossDetectorTest
         }
     }
     private static final int MESSAGE_LENGTH = DataHeaderFlyweight.HEADER_LENGTH + DATA.length;
-    private static final int ALIGNED_FRAME_LENGTH = (int) align(MESSAGE_LENGTH, FrameDescriptor.FRAME_ALIGNMENT);
+    private static final int ALIGNED_FRAME_LENGTH = align(MESSAGE_LENGTH, FrameDescriptor.FRAME_ALIGNMENT);
     private static final int SESSION_ID = 0x5E55101D;
     private static final int STREAM_ID = 0xC400E;
     private static final int TERM_ID = 0xEE81D;
