+++ aeron-driver/src/test/java/uk/co/real_logic/aeron/driver/RetransmitHandlerTest.java
@@ -45,7 +45,7 @@ public class RetransmitHandlerTest
     private static final int META_DATA_BUFFER_LENGTH = LogBufferDescriptor.TERM_META_DATA_LENGTH;
     private static final byte[] DATA = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15};
     private static final int MESSAGE_LENGTH = DataHeaderFlyweight.HEADER_LENGTH + DATA.length;
-    private static final int ALIGNED_FRAME_LENGTH = (int) align(MESSAGE_LENGTH, FrameDescriptor.FRAME_ALIGNMENT);
+    private static final int ALIGNED_FRAME_LENGTH = align(MESSAGE_LENGTH, FrameDescriptor.FRAME_ALIGNMENT);
     private static final int SESSION_ID = 0x5E55101D;
     private static final int STREAM_ID = 0x5400E;
     private static final int TERM_ID = 0x7F003355;
